This is an attempt to implement a turn-based game when the two players use distinct forms to interact with the world state, e.g.,
 -- one reacts on tick events
 -- the other reacts to mouse events and/or key strokes.

My solution is brittle (introducing turns, using game-specific knowledge);
perhaps there is a better one ..

Once I had the solution, it turns out that the game might not be
termination if the other player is forced to eat all enemies. There
will be situations when no strategy works (perhaps the step-size
is too large. I need to check.)