"""    A human player can lose points using just the keyboard by trying to    destroy an enemy (with the space bar) when his fighter isn't on the    red or blue planet. --- This cannot happen when the player    accidentally uses the mouse and just clicks on a planet. """
In a sense, this is a bug. I forgot the data interpretation of the
`eat` action -- to eat the 'pill' on which the fighter currently rests
and passed in the position of the 'clicked-on' pill instead. While
the fighter couldn't 'eat' it, `execute` also didn't subtract points
for the bad attempt at eating a pill. -- Straightforward fix.

Contract: a contract on `eat` that knows the state could have avoided
the problem.

Design: in a sense the eat action doesn't need an argument at all.
The 'referee' knows that the first fighter is the one that is trying to eat a pill.

BUT, I just realized that my seemingly straightforward fix introduced
another bug: now a player can click on a far-away pill while the
fighter sits on a different one .. and the eat action will succceed.